requirements.txt
# Core
streamlit>=1.30.0
python-dotenv>=1.0.0

# Joke Generator (Version_12)
google-genai>=1.0.0
openai>=1.0.0
supabase>=2.0.0

# Video Studio (Studio_AutoPost)
moviepy>=2.2.1
Pillow>=11.3.0
numpy>=1.25
requests>=2.31.0

# Twitter/X API
requests-oauthlib>=1.3.1

# Scheduler
pytz>=2023.3
python-dateutil>=2.8.0

# News Workflow (Daily Comedy Brief)
gnews>=0.3.0
